data-generator/requirements.txt
psycopg2-binary
pyyaml
faker